% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_filter.R
\name{cell_filter}
\alias{cell_filter}
\title{Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.}
\usage{
cell_filter(summary, nUMI.thresh, det.rate.thresh, percent.mito.thresh)
}
\arguments{
\item{nUMI.thresh}{: numeric vector of length 2 for lower and upper bound for nUMI}

\item{det.rate.thresh}{: numeric vector of length 2 for lower and upper bound for detection rate}

\item{percent.mito.thresh}{: numeric vector of length 2 for lower and upper bound for the proportion of mito-genes}

\item{X}{log-transformed expression level matrix with cells at columns, genes at rows}
}
\description{
Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.
}
