% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellFilter.R
\name{cellFilter}
\alias{cellFilter}
\title{Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.}
\usage{
cellFilter(X, genenames, minGene = -Inf, maxGene = Inf,
  maxMitoProp = 0.1)
}
\arguments{
\item{X}{log-transformed expression level matrix with cells at columns, genes at rows}

\item{genenames}{: gene names for each row of X}

\item{minGene}{: lower bound of detected genes of a cell}

\item{maxGene}{: upper bound of detected genes of a cell}

\item{maxMitoProp}{: maximum possible value of the proportion of mito-gene reads per cell.}
}
\description{
Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.
}
