% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_detection_rate.R
\name{correct_detection_rate}
\alias{correct_detection_rate}
\title{Returns scatterplot of detection rate and the first pc of the log-transformed expression level matrix
and the matrix with the detection rate regressed out
Users can decide whether to use the raw matrix or the residual matrix based on the visualization}
\usage{
correct_detection_rate(X, det.rate)
}
\arguments{
\item{X}{log-transformed expression level matrix with cells at columns, genes at rows}

\item{det.rate}{detection rate : usually colSums(X>0) / nrow(X) where X is count matrix.}
}
\description{
Returns scatterplot of detection rate and the first pc of the log-transformed expression level matrix
and the matrix with the detection rate regressed out
Users can decide whether to use the raw matrix or the residual matrix based on the visualization
}
\examples{

X = SCNoisyClustering::yan
det.rate = colSums(X>0) / nrow(X)
out = correct_detection_rate(X, det.rate)
out$plot #check if there's a distinct relationship
X = out$residual #if you decide to regress out the detection rate

}
