% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_genes.R
\name{de_genes}
\alias{de_genes}
\title{Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.}
\usage{
de_genes(X, genenames, cluster, top.n = 100, plot = 6)
}
\arguments{
\item{X}{log-transformed expression level matrix with cells at columns, genes at rows}

\item{mean.thresh}{: numeric vector of length 2 for lower and upper bound for the mean expression level}

\item{dispersion.thresh}{: numeric vector of length 2 for lower and upper bound for log variance to mean ratio}
}
\value{
expression level matrix containing only highly variable genes
}
\description{
Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.
}
