% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.data.R
\name{explore.data}
\alias{explore.data}
\title{Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.}
\usage{
explore.data(X, genenames)
}
\arguments{
\item{X}{log-transformed expression level matrix with cells at columns, genes at rows}

\item{genenames}{: gene names for each row of X}
}
\description{
Remove abnormal cells and return the summary of each cell including the UMI count, gene count, percent of mito-gene, and detection rate.
}
