% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genefilter.R
\name{genefilter}
\alias{genefilter}
\title{Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.}
\usage{
genefilter(X, p1 = 0.9, p2 = 0)
}
\arguments{
\item{X}{data matrix with zeros}

\item{p1}{upper threshold of the percentage of zeros in each row}

\item{p2}{lower threshold of the percentage of zeros in each row}
}
\value{
The data matrix with rows with too many or too few zeros removed
}
\description{
Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.
}
\examples{
X = matrix(sample(0:10, size = 10000, replace = TRUE, prob = c(0.9, rep(0.1/10, 10))), nrow = 200) #create expression level matrix
newX = genefilter(X, 0.9, 0)
dim(newX)
}
