% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClustNum.R
\name{getClustNum}
\alias{getClustNum}
\title{Estimate the optimal number of clusters using eigengap}
\usage{
getClustNum(S)
}
\arguments{
\item{S}{estimated similarity matrix.}
}
\description{
Estimate the optimal number of clusters using eigengap
}
\examples{
#create positive definite symmetric matrix
X = matrix(rnorm(50), nrow = 10)
S= t(X) \%*\% X
getClustNum(S)
}
