% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{network.diffusion}
\alias{network.diffusion}
\title{Perform network diffusion}
\usage{
network.diffusion(S, K, alpha = 0.7)
}
\arguments{
\item{S}{Estimated similarity matrix.}

\item{alpha}{diffusion parameter}

\item{k}{Number of neighbors for knn}
}
\description{
Perform network diffusion
}
