% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dispersion.R, R/remove.zero.genes.R
\name{plot.dispersion}
\alias{plot.dispersion}
\title{Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.}
\usage{
plot.dispersion(X, bins = NA, median = FALSE,
  outliers.mean.thresh = c(30, Inf), outliers.vmr.thresh = c(3, Inf))

plot.dispersion(X, bins = NA, median = FALSE,
  outliers.mean.thresh = c(30, Inf), outliers.vmr.thresh = c(3, Inf))
}
\arguments{
\item{X}{data matrix with zeros}

\item{bins}{number of bins to divide genes by their mean expression level}

\item{X}{data matrix with zeros}

\item{bins}{number of bins to divide genes by their mean expression level}
}
\value{
data frame with mean and variance of expression level for each gene, bins they belong to, and normalized dispersion level

data frame with mean and variance of expression level for each gene, bins they belong to, and normalized dispersion level
}
\description{
Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.

Remove genes (rows) with little information.
Genes that are not expressed in any of the samples, or most of the samples, and in some cases, genes that are expressed in most the samples, will be removed.
}
\examples{
X = matrix(sample(0:10, size = 10000, replace = TRUE, prob = c(0.9, rep(0.1/10, 10))), nrow = 200) #create expression level matrix
newX = genefilter(X, 0.9, 0)
dim(newX)
}
