% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dispersion.R
\name{plot_dispersion}
\alias{plot_dispersion}
\title{Show distribution of normalized dispersion for each gene}
\usage{
plot_dispersion(X, genenames, bins = NA, median = FALSE,
  outliers.mean.thresh = c(30, Inf), outliers.vmr.thresh = c(3, Inf))
}
\arguments{
\item{X}{data matrix}

\item{genenames}{name of the genes}

\item{bins}{number of bins to divide genes by their mean expression level}

\item{median}{whether to use median instead of mean, MAD instead of SD}

\item{outliers.mean.thresh}{x-axis threshold to show outliergene names instead of simple plotting}

\item{outliers.vmr.thresh}{y-axis threshold to show outlier gene names instead of simple plotting}
}
\description{
Show distribution of normalized dispersion for each gene
}
