% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.normalize.R
\name{quantile.normalize}
\alias{quantile.normalize}
\title{quantile normalize the gene expression level matrix}
\usage{
\method{quantile}{normalize}(X)
}
\arguments{
\item{X}{matrix to quantile-normalize}
}
\description{
quantile normalize the gene expression level matrix
}
