% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_normalize.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{quantile normalize the gene expression level matrix}
\usage{
quantile_normalize(X)
}
\arguments{
\item{X}{matrix to quantile-normalize}
}
\description{
quantile normalize the gene expression level matrix
}
