% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_sparse_kernel.R
\name{save_sparse_kernel}
\alias{save_sparse_kernel}
\title{For large version, save the large kernel matrices in the hard drive}
\usage{
save_sparse_kernel(X, dir = ".", kernel_type = kernel_type,
  klist = seq(15, 25, by = 5), sigmalist = seq(1, 2, by = 0.2),
  verbose = FALSE)
}
\arguments{
\item{X}{Data matrix to construct kernel from}

\item{dir}{Directory to save the matrix, default as the current directory}

\item{kernel_type}{Type of distance measure to use : euclidean, pearson, spearman, or combined}

\item{klist}{Kernel parameters k}

\item{sigmalist}{Kernel parameters sigma}

\item{verbose}{Show progress if set to TRUE}
}
\description{
For large version, save the large kernel matrices in the hard drive
}
