% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SCNoisyClustering}
\alias{SCNoisyClustering}
\alias{SCNoisyClustering-package}
\title{SCNoisyClustering
Single Cell Clustering}
\description{
SCNoisyClustering
Single Cell Clustering
}
\author{
Tae Kim <tk382@uchicago.edu>
}
