% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCscale.R
\name{SCscale}
\alias{SCscale}
\title{Scale each gene to have mean 0 and variance 1}
\usage{
SCscale(X)
}
\arguments{
\item{X}{log-transformed expression level matrix}
}
\description{
Scale each gene to have mean 0 and variance 1
}
