% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_spectral.R
\name{tsne_spectral}
\alias{tsne_spectral}
\title{dimension reduction wrapper}
\usage{
tsne_spectral(A, numClust, numEigen = NA)
}
\arguments{
\item{A}{Estimated similarity matrix.}

\item{numClust}{Number of clusters. This can't be NA. If unknown, it should be estimated from getNumClust function using A}

\item{numEigen}{number of eigenvectors to be used. If NA, the number of clusters will be used}
}
\description{
dimension reduction wrapper
}
\examples{

}
