% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI.normalize.R
\name{UMI_normalize}
\alias{UMI_normalize}
\title{Scale each gene to have mean 0 and variance 1}
\usage{
UMI_normalize(X)
}
\arguments{
\item{X}{Raw count matrix}
}
\description{
Scale each gene to have mean 0 and variance 1
}
