% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesparsekernel.R
\name{makesparsekernel}
\alias{makesparsekernel}
\title{For large version, save the large kernel matrices in the hard drive}
\usage{
makesparsekernel(X, kernel_type = kernel_type, klist = seq(15, 25, by =
  5), sigmalist = seq(1, 2, by = 0.5), verbose = FALSE)
}
\arguments{
\item{X}{Data matrix to construct kernel from}

\item{kernel_type}{Type of distance measure to use : euclidean, pearson, spearman, or combined}

\item{klist}{Kernel parameters k}

\item{sigmalist}{Kernel parameters sigma}

\item{verbose}{Show progress if set to TRUE}
}
\description{
For large version, save the large kernel matrices in the hard drive
}
