% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{network.diffusion}
\alias{network.diffusion}
\title{Estimate the optimal number of clusters using eigengap}
\usage{
network.diffusion(A, K, alpha = 0.7)
}
\arguments{
\item{S}{estimated similarity matrix.}
}
\description{
Estimate the optimal number of clusters using eigengap
}
\examples{
#create positive definite symmetric matrix
X = matrix(rnorm(50), nrow = 10)
S= t(X) \%*\% X
getClustNum(S)
}
