% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savesparsekernel.R
\name{savesparsekernel}
\alias{savesparsekernel}
\title{Estimate the optimal number of clusters using eigengap}
\usage{
savesparsekernel(X, dir = ".", kernel_type = kernel_type, klist = seq(15,
  25, by = 5), sigmalist = seq(1, 2, by = 0.2), verbose = FALSE)
}
\arguments{
\item{S}{estimated similarity matrix.}
}
\description{
Estimate the optimal number of clusters using eigengap
}
\examples{
#create positive definite symmetric matrix
X = matrix(rnorm(50), nrow = 10)
S= t(X) \%*\% X
getClustNum(S)
}
