% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLSL.R
\name{SLSL}
\alias{SLSL}
\title{Wrapper function for similarity learning algorithm. For details, refer to the manuscript.
We recommend users to start with default setting for filter, kernel_type, klist, sigmalist, tau, gamma, and k.
The only critical parameter is log. If X has been log-transformed and log is not set to FALSE, it will produce an error.}
\usage{
SLSL(X, plot = TRUE, numClust = NA, kernel_type = "combined",
  klist = NA, sigmalist = seq(1, 2, by = 0.5), tau = 5, gamma = 0,
  verbose = FALSE, warning = TRUE)
}
\arguments{
\item{X}{Expression level matrix in count, cells in columns, genes in rows}

\item{plot}{if set to TRUE, shows the plot of the dimension reduction result}

\item{numClust}{Number of Clusters. If unknown, NA. If set to NA, it will be estimated with minimum 3.}

\item{kernel_type}{distance measure to use: one of "euclidean", "pearson", "spearman", or "combined"}

\item{klist}{kernel parameters setting}

\item{sigmalist}{kernel parameters setting.}

\item{tau}{Regularization parameter for L1 norm}

\item{gamma}{Regularization parameter for Frobenius norm}

\item{verbose}{To show progress, set to TRUE}
}
\description{
Wrapper function for similarity learning algorithm. For details, refer to the manuscript.
We recommend users to start with default setting for filter, kernel_type, klist, sigmalist, tau, gamma, and k.
The only critical parameter is log. If X has been log-transformed and log is not set to FALSE, it will produce an error.
}
